Update parser.py
After each iteration of loop, new frame is appended to existing frame. Then, the combined existing frame is written to Excel sheet with appropriate sheet name. After loop finishes, the average and minimum calculations are performed and a third data frame is created. The calculation data frame is appended to the existing data frame and the combined data frame is finally written to the Excel file with appropriate sheet name.